requirements.txt
numpy
scipy
nibabel
dipy
pillow
